Refactor and update drawing and processing services
Updated Job.csproj to include new services and remove old ones. Consolidated DrawPageFront and DrawPageBack into DrawPage in Proof.cs. Updated Back.cs, Front.cs, and WorkAndTurn.cs to use the new DrawPage method. Replaced DrawSingleSideService and DrawWorkAndTurnService with ScreenDrawSingleSideService and ScreenDrawWorkAndTurnService in ScreenDrawer.cs. Made minor formatting adjustments in Front.cs and refactored GetMediabox method. Removed DrawSingleSideService.cs and ProcessCenterH.cs, ProcessCenterV.cs, moving their functionality to ProcessMoveSubject.cs. Updated PdfMarksService and TextMarksService to use ProcessSubject.GetSubjectRect. Updated FormPdfImposition and FormAddTemplatePlate to use new centering methods. Refactored PreviewControl to use ScreenDrawCommons. Added ScreenDrawCommons.cs with methods for drawing and calculating dimensions. Cleaned up build configurations in work.sln. Added ProcessFixBleeds.cs and ProcessSubject.cs with methods for adjusting bleeds and calculating subject rectangles.
