+++ Job/Static/Pdf/Imposition/Models/TemplatePage.cs
@@ -1,4 +1,5 @@
-﻿using JobSpace.Static.Pdf.Imposition.Services;
+﻿using JobSpace.Static.Pdf.Imposition.Drawers.Services.Screen;
+using JobSpace.Static.Pdf.Imposition.Services;
 using JobSpace.Static.Pdf.Imposition.Services.Impos.Binding.Loose.Sheetwise;
 using System;
 using System.Collections.Generic;
@@ -86,55 +87,6 @@ namespace JobSpace.Static.Pdf.Imposition.Models
             return GetClippedW;
         }
 
-        public double GetPageDrawFrontX()
-        {
-            switch (Front.Angle)
-            {
-                case 0:
-                    return Front.X + Margins.Left;
-                case 90:
-                    return Front.X + Margins.Top;
-                case 180:
-                    return Front.X + Margins.Bottom;
-                case 270:
-                    return Front.X + Margins.Right;
-                default:
-                    throw new NotImplementedException();
-            }
-        }
-
-
-
-        public double GetPageDrawFrontY()
-        {
-
-            switch (Front.Angle)
-            {
-                case 0:
-                    break;
-                case 90:
-                    return Front.Y + Margins.Right;
-                case 180:
-                    return Front.Y + Margins.Top;
-                case 270:
-                    return Front.Y + Margins.Left;
-
-            }
-            return Front.Y + Margins.Bottom;
-        }
-
-        public double GetPageDrawFrontW()
-        {
-            if (Front.Angle == 0 || Front.Angle == 180) return W;
-            return H;
-        }
-
-        public double GetPageDrawFrontH()
-        {
-            if (Front.Angle == 0 || Front.Angle == 180) return H;
-            return W;
-        }
-
         public void FlipAngle(TemplateSheetPlaceType sheetPlaceType)
         {
             Front.Angle = (Front.Angle + 180) % 360;
@@ -167,167 +119,8 @@ namespace JobSpace.Static.Pdf.Imposition.Models
             return JsonSerializer.Deserialize<TemplatePage>(str);
         }
 
-        public RectangleD GetDrawBleedFrontLeft()
-        {
-            switch (Front.Angle)
-            {
-                case 0:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X - Bleeds.Left,
-                        Y1 = Front.Y,
-                        X2 = Front.X,
-                        Y2 = Front.Y + GetPageDrawFrontH(),
-                    }
-                        ;
-                case 90:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X,
-                        Y1 = Front.Y - Bleeds.Left,
-                        X2 = Front.X + GetPageDrawFrontW(),
-                        Y2 = Front.Y
-                    };
-                case 180:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X + GetPageDrawFrontW(),
-                        Y1 = Front.Y,
-                        X2 = Front.X + GetPageDrawFrontW() + Bleeds.Left,
-                        Y2 = Front.Y + GetPageDrawFrontH(),
-                    };
-                case 270:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X,
-                        Y1 = Front.Y + GetPageDrawFrontH(),
-                        X2 = Front.X + GetPageDrawFrontW(),
-                        Y2 = Front.Y + GetPageDrawFrontH() + Bleeds.Left,
-                    };
-                default:
-                    throw new NotImplementedException();
-            }
-        }
-        public RectangleD GetDrawBleedFrontRight()
-        {
-            switch (Front.Angle)
-            {
-                case 0:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X + GetPageDrawFrontW(),
-                        Y1 = Front.Y,
-                        X2 = Front.X + GetPageDrawFrontW() + Bleeds.Right,
-                        Y2 = Front.Y + GetPageDrawFrontH(),
-                    };
-                case 90:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X,
-                        Y1 = Front.Y + GetPageDrawFrontH(),
-                        X2 = Front.X + GetPageDrawFrontW(),
-                        Y2 = Front.Y + GetPageDrawFrontH() + Bleeds.Right
-                    };
-                case 180:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X - Bleeds.Right,
-                        Y1 = Front.Y,
-                        X2 = Front.X,
-                        Y2 = Front.Y + GetPageDrawFrontH()
-                    };
-                case 270:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X,
-                        Y1 = Front.Y - Bleeds.Right,
-                        X2 = Front.X + GetPageDrawFrontW(),
-                        Y2 = Front.Y
-                    };
-                default:
-                    throw new NotImplementedException();
-            }
-        }
-        public RectangleD GetDrawBleedFrontBottom()
-        {
-            switch (Front.Angle)
-            {
-                case 0:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X,
-                        Y1 = Front.Y - Bleeds.Bottom,
-                        X2 = Front.X + GetPageDrawFrontW(),
-                        Y2 = Front.Y
-                    };
-                case 90:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X + GetPageDrawFrontW(),
-                        Y1 = Front.Y,
-                        X2 = Front.X + GetPageDrawFrontW() + Bleeds.Bottom,
-                        Y2 = Front.Y + GetPageDrawFrontH()
-                    };
-                case 180:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X,
-                        Y1 = Front.Y + GetPageDrawFrontH(),
-                        X2 = Front.X + GetPageDrawFrontW(),
-                        Y2 = Front.Y + GetPageDrawFrontH() + Bleeds.Bottom,
-                    };
-                case 270:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X - Bleeds.Bottom,
-                        Y1 = Front.Y,
-                        X2 = Front.X,
-                        Y2 = Front.Y + GetPageDrawFrontH()
-                    };
-                default:
-                    throw new NotImplementedException();
-            }
-        }
-        public RectangleD GetDrawBleedFrontTop()
-        {
-            switch (Front.Angle)
-            {
-                case 0:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X,
-                        Y1 = Front.Y + GetPageDrawFrontH(),
-                        X2 = Front.X + GetPageDrawFrontW(),
-                        Y2 = Front.Y + GetPageDrawFrontH() + Bleeds.Top
-                    };
-                case 90:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X - Bleeds.Top,
-                        Y1 = Front.Y,
-                        X2 = Front.X,
-                        Y2 = Front.Y + GetPageDrawFrontH()
-                    };
-                case 180:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X,
-                        Y1 = Front.Y - Bleeds.Top,
-                        X2 = Front.X + GetPageDrawFrontW(),
-                        Y2 = Front.Y
-                    };
-                case 270:
-                    return new RectangleD()
-                    {
-                        X1 = Front.X + GetPageDrawFrontW(),
-                        Y1 = Front.Y,
-                        X2 = Front.X + GetPageDrawFrontW() + Bleeds.Top,
-                        Y2 = Front.Y + GetPageDrawFrontH()
-                    };
-                default:
-                    throw new NotImplementedException();
-            }
-        }
+        
+      
 
         public void SwitchWH()
         {
