+++ Job/Static/Pdf/Imposition/Services/Impos/Binding/Loose/Sheetwise/LooseBindingSheetwise.cs
@@ -1,4 +1,5 @@
-﻿using JobSpace.Static.Pdf.Imposition.Models;
+﻿using JobSpace.Static.Pdf.Imposition.Drawers.Services.Screen;
+using JobSpace.Static.Pdf.Imposition.Models;
 using JobSpace.Static.Pdf.Imposition.Services.Impos.Models;
 using System;
 using System.Collections.Generic;
@@ -115,62 +116,7 @@ namespace JobSpace.Static.Pdf.Imposition.Services.Impos.Binding.Loose.Sheetwise
             }
         }
 
-        public static void FixBleedsBack(TemplatePageContainer templatePageContainer)
-        {
-            templatePageContainer.TemplatePages.ForEach(x => x.Bleeds.Set(x.Bleeds.Default));
-
-            foreach (var page in templatePageContainer.TemplatePages)
-            {
-                RectangleD left = page.GetDrawBleedFrontLeft();
-                RectangleD right = page.GetDrawBleedFrontRight();
-                RectangleD top = page.GetDrawBleedFrontTop();
-                RectangleD bottom = page.GetDrawBleedFrontBottom();
-
-                foreach (var pageTarget in templatePageContainer.TemplatePages)
-                {
-                    if (page != pageTarget)
-                    {
-                        RectangleD pageRect = new RectangleD
-                        {
-                            X1 = pageTarget.GetPageDrawFrontX(),
-                            Y1 = pageTarget.GetPageDrawFrontY(),
-                            X2 = pageTarget.GetPageDrawFrontX() + pageTarget.GetPageDrawFrontW(),
-                            Y2 = pageTarget.GetPageDrawFrontY() + pageTarget.GetPageDrawFrontH()
-                        };
-
-                        List<RectangleD> rects = new List<RectangleD>(){
-                            pageTarget.GetDrawBleedFrontLeft(),
-                            pageTarget.GetDrawBleedFrontRight(),
-                            pageTarget.GetDrawBleedFrontTop(),
-                            pageTarget.GetDrawBleedFrontBottom()
-                        };
-
-                        foreach (var rect in rects)
-                        {
-                            if (left.IntersectsWith(rect) || left.IntersectsWith(pageRect))
-                            {
-                                page.Bleeds.Left = 0;
-
-                            }
-                            if (right.IntersectsWith(rect) || right.IntersectsWith(pageRect))
-                            {
-                                page.Bleeds.Right = 0;
-
-                            }
-                            if (top.IntersectsWith(rect) || top.IntersectsWith(pageRect))
-                            {
-                                page.Bleeds.Top = 0;
-                            }
-                            if (bottom.IntersectsWith(rect) || bottom.IntersectsWith(pageRect))
-                            {
-                                page.Bleeds.Bottom = 0;
-                            }
-                        }
-                    }
-                }
-
-            }
-        }
+        
 
         public static void FixBackPagePosition(TemplateSheet sheet, TemplatePage selectedPreviewPage)
         {
