+++ Job/Static/Pdf/Imposition/Services/Impos/Binding/Loose/SingleSide/LooseBindingSingleSide.cs
@@ -1,4 +1,6 @@
-﻿using JobSpace.Static.Pdf.Imposition.Models;
+﻿using JobSpace.Static.Pdf.Imposition.Drawers.Services.Screen;
+using JobSpace.Static.Pdf.Imposition.Models;
+using JobSpace.Static.Pdf.Imposition.Services.Impos.Processes;
 using System;
 using System.Collections.Generic;
 
@@ -177,7 +179,7 @@ namespace JobSpace.Static.Pdf.Imposition.Services.Impos
         {
             if (parameters.IsOneCut)
             {
-                FixBleedsFront(templatePageContainer);
+                ProcessFixBleeds.Front(templatePageContainer);
             }
 
             CropMarksService.FixCropMarksFront(templatePageContainer);
@@ -224,59 +226,61 @@ namespace JobSpace.Static.Pdf.Imposition.Services.Impos
             return (sheetW, sheetH);
         }
 
-        public static void FixBleedsFront(TemplatePageContainer templatePageContainer)
-        {
-            templatePageContainer.TemplatePages.ForEach(x => x.Bleeds.Set(x.Bleeds.Default));
+        //public static void FixBleedsFront(TemplatePageContainer templatePageContainer)
+        //{
+        //    templatePageContainer.TemplatePages.ForEach(x => x.Bleeds.Set(x.Bleeds.Default));
 
-            foreach (var page in templatePageContainer.TemplatePages)
-            {
-                RectangleD left = page.GetDrawBleedFrontLeft();
-                RectangleD right = page.GetDrawBleedFrontRight();
-                RectangleD top = page.GetDrawBleedFrontTop();
-                RectangleD bottom = page.GetDrawBleedFrontBottom();
+        //    foreach (var page in templatePageContainer.TemplatePages)
+        //    {
+        //        PageSide side = page.Front;
 
-                foreach (var pageTarget in templatePageContainer.TemplatePages)
-                {
-                    if (page != pageTarget)
-                    {
-                        RectangleD pageRect = new RectangleD
-                        {
-                            X1 = pageTarget.GetPageDrawFrontX(),
-                            Y1 = pageTarget.GetPageDrawFrontY(),
-                            X2 = pageTarget.GetPageDrawFrontX() + pageTarget.GetPageDrawFrontW(),
-                            Y2 = pageTarget.GetPageDrawFrontY() + pageTarget.GetPageDrawFrontH()
-                        };
+        //        RectangleD left = page.GetDrawBleedFrontLeft();
+        //        RectangleD right = page.GetDrawBleedFrontRight();
+        //        RectangleD top = page.GetDrawBleedFrontTop();
+        //        RectangleD bottom = page.GetDrawBleedFrontBottom();
 
-                        List<RectangleD> rects = new List<RectangleD>() {
-                            pageTarget.GetDrawBleedFrontLeft(),
-                            pageTarget.GetDrawBleedFrontRight(),
-                            pageTarget.GetDrawBleedFrontTop(),
-                            pageTarget.GetDrawBleedFrontBottom()
-                        };
+        //        foreach (var pageTarget in templatePageContainer.TemplatePages)
+        //        {
+        //            if (page != pageTarget)
+        //            {
+        //                RectangleD pageRect = new RectangleD
+        //                {
+        //                    X1 = ScreenDrawCommons.GetPageDrawX(page,side),
+        //                    Y1 = ScreenDrawCommons.GetPageDrawY(page, side),
+        //                    X2 = ScreenDrawCommons.GetPageDrawX(page,side) + ScreenDrawCommons.GetPageDrawW(page,side),
+        //                    Y2 = ScreenDrawCommons.GetPageDrawY(page,side) + ScreenDrawCommons.GetPageDrawH(page,side)
+        //                };
 
-                        foreach (var rect in rects)
-                        {
-                            if (left.IntersectsWith(rect) || left.IntersectsWith(pageRect))
-                            {
-                                page.Bleeds.Left = 0;
-                            }
-                            if (right.IntersectsWith(rect) || right.IntersectsWith(pageRect))
-                            {
-                                page.Bleeds.Right = 0;
-                            }
-                            if (top.IntersectsWith(rect) || top.IntersectsWith(pageRect))
-                            {
-                                page.Bleeds.Top = 0;
-                            }
-                            if (bottom.IntersectsWith(rect) || bottom.IntersectsWith(pageRect))
-                            {
-                                page.Bleeds.Bottom = 0;
-                            }
-                        }
-                    }
-                }
-            }
-        }
+        //                List<RectangleD> rects = new List<RectangleD>() {
+        //                    pageTarget.GetDrawBleedFrontLeft(),
+        //                    pageTarget.GetDrawBleedFrontRight(),
+        //                    pageTarget.GetDrawBleedFrontTop(),
+        //                    pageTarget.GetDrawBleedFrontBottom()
+        //                };
+
+        //                foreach (var rect in rects)
+        //                {
+        //                    if (left.IntersectsWith(rect) || left.IntersectsWith(pageRect))
+        //                    {
+        //                        page.Bleeds.Left = 0;
+        //                    }
+        //                    if (right.IntersectsWith(rect) || right.IntersectsWith(pageRect))
+        //                    {
+        //                        page.Bleeds.Right = 0;
+        //                    }
+        //                    if (top.IntersectsWith(rect) || top.IntersectsWith(pageRect))
+        //                    {
+        //                        page.Bleeds.Top = 0;
+        //                    }
+        //                    if (bottom.IntersectsWith(rect) || bottom.IntersectsWith(pageRect))
+        //                    {
+        //                        page.Bleeds.Bottom = 0;
+        //                    }
+        //                }
+        //            }
+        //        }
+        //    }
+        //}
