+++ Job/Static/Pdf/Imposition/Services/Impos/Processes/ProcessMoveSubject.cs
@@ -43,5 +43,58 @@ namespace JobSpace.Static.Pdf.Imposition.Services.Impos.Processes
                 ProcessFixPageBackPosition.FixPosition(sheet, item);
             }
         }
+
+        public static void CenterVer(TemplateSheet sheet)
+        {
+            if (sheet?.TemplatePageContainer?.TemplatePages == null || !sheet.TemplatePageContainer.TemplatePages.Any())
+                return;
+
+            var subjectRect = ProcessSubject.GetSubjectRect(sheet, sheet.TemplatePageContainer);
+            var printRect = ProcessSubject.GetSubjectRect(sheet);
+
+            // Calculate the vertical offset needed to center the content
+            var verticalOffset = (printRect.H - subjectRect.H) / 2;
+            var adjustedY = subjectRect.Y1 - verticalOffset;
+
+            // Adjust the Y position of each template page
+            foreach (var page in sheet.TemplatePageContainer.TemplatePages)
+            {
+                page.Front.Y -= (adjustedY - sheet.SafeFields.Bottom);
+                ProcessFixPageBackPosition.FixPosition(sheet, page);
+            }
+        }
+
+        public static void CenterHor(TemplateSheet sheet)
+        {
+            if (sheet?.TemplatePageContainer?.TemplatePages == null || !sheet.TemplatePageContainer.TemplatePages.Any())
+                return;
+
+            try
+            {
+                // Calculate the bounding rectangle of the pages
+                var subjectRect = ProcessSubject.GetSubjectRect(sheet, sheet.TemplatePageContainer);
+
+                // Get the printable area of the sheet
+                var printRect = ProcessSubject.GetSubjectRect(sheet);
+
+                // Calculate the horizontal offset needed to center the content
+                var horizontalOffset = (printRect.W - subjectRect.W) / 2;
+
+                // Calculate the new X offset
+                var newXOffset = subjectRect.X1 - horizontalOffset;
+
+                // Adjust the X position of each page to center them
+                foreach (var page in sheet.TemplatePageContainer.TemplatePages)
+                {
+                    page.Front.X -= newXOffset - sheet.SafeFields.Left;
+                    ProcessFixPageBackPosition.FixPosition(sheet, page);
+                }
+            }
+            catch (Exception ex)
+            {
+                // Log or handle the exception as needed
+                Console.WriteLine($"An error occurred while centering the sheet: {ex.Message}");
+            }
+        }
     }
 }
