+++ work.sln
@@ -139,7 +139,6 @@ Global
 		{20640AA0-AB38-4499-8B96-99143A3B033E}.SERVER|x64.Build.0 = Release|x64
 		{20640AA0-AB38-4499-8B96-99143A3B033E}.SERVER|x86.ActiveCfg = Release|Any CPU
 		{CD75770A-E81D-49E6-9E9C-CF104172C48F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CD75770A-E81D-49E6-9E9C-CF104172C48F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CD75770A-E81D-49E6-9E9C-CF104172C48F}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{CD75770A-E81D-49E6-9E9C-CF104172C48F}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{CD75770A-E81D-49E6-9E9C-CF104172C48F}.Debug|x64.ActiveCfg = Debug|x64
@@ -184,7 +183,6 @@ Global
 		{32FF055B-5267-4802-BDC7-94D965CD730C}.SERVER|x64.Build.0 = Release|x64
 		{32FF055B-5267-4802-BDC7-94D965CD730C}.SERVER|x86.ActiveCfg = Release|Any CPU
 		{348B842C-D367-46BC-9C46-4E5114EE9B56}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{348B842C-D367-46BC-9C46-4E5114EE9B56}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{348B842C-D367-46BC-9C46-4E5114EE9B56}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{348B842C-D367-46BC-9C46-4E5114EE9B56}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{348B842C-D367-46BC-9C46-4E5114EE9B56}.Debug|x64.ActiveCfg = Debug|x64
@@ -230,7 +228,6 @@ Global
 		{905750A5-F025-47BB-9F5A-81D7EBD0277D}.SERVER|x86.ActiveCfg = Release|Any CPU
 		{905750A5-F025-47BB-9F5A-81D7EBD0277D}.SERVER|x86.Build.0 = Release|Any CPU
 		{1293B563-EEA2-429F-8866-8DE86B2D1D6A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1293B563-EEA2-429F-8866-8DE86B2D1D6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1293B563-EEA2-429F-8866-8DE86B2D1D6A}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{1293B563-EEA2-429F-8866-8DE86B2D1D6A}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{1293B563-EEA2-429F-8866-8DE86B2D1D6A}.Debug|x64.ActiveCfg = Debug|x64
@@ -253,7 +250,6 @@ Global
 		{1293B563-EEA2-429F-8866-8DE86B2D1D6A}.SERVER|x86.ActiveCfg = Release|Any CPU
 		{1293B563-EEA2-429F-8866-8DE86B2D1D6A}.SERVER|x86.Build.0 = Release|Any CPU
 		{EE3D7AAA-F7A1-4B37-981A-848A4A6831D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EE3D7AAA-F7A1-4B37-981A-848A4A6831D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EE3D7AAA-F7A1-4B37-981A-848A4A6831D7}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{EE3D7AAA-F7A1-4B37-981A-848A4A6831D7}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{EE3D7AAA-F7A1-4B37-981A-848A4A6831D7}.Debug|x64.ActiveCfg = Debug|x64
@@ -321,7 +317,6 @@ Global
 		{CF07395A-59A1-4244-9AAE-11F1ACD74747}.SERVER|x86.ActiveCfg = Release|Any CPU
 		{CF07395A-59A1-4244-9AAE-11F1ACD74747}.SERVER|x86.Build.0 = Release|Any CPU
 		{BB0FE1FA-8C37-4E59-A548-669F97109E64}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BB0FE1FA-8C37-4E59-A548-669F97109E64}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BB0FE1FA-8C37-4E59-A548-669F97109E64}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{BB0FE1FA-8C37-4E59-A548-669F97109E64}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{BB0FE1FA-8C37-4E59-A548-669F97109E64}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -389,7 +384,6 @@ Global
 		{C3715EAD-232D-4B0F-8D21-0D50BF2CA4CA}.SERVER|x86.ActiveCfg = Release|Any CPU
 		{C3715EAD-232D-4B0F-8D21-0D50BF2CA4CA}.SERVER|x86.Build.0 = Release|Any CPU
 		{AC6CFED7-8E64-469F-A97D-C25B644CD8A3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AC6CFED7-8E64-469F-A97D-C25B644CD8A3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AC6CFED7-8E64-469F-A97D-C25B644CD8A3}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{AC6CFED7-8E64-469F-A97D-C25B644CD8A3}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{AC6CFED7-8E64-469F-A97D-C25B644CD8A3}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -436,7 +430,6 @@ Global
 		{AF13E994-0F4D-42AF-A20F-6D3ABD53F9FE}.SERVER|x86.ActiveCfg = Release|Any CPU
 		{AF13E994-0F4D-42AF-A20F-6D3ABD53F9FE}.SERVER|x86.Build.0 = Release|Any CPU
 		{75C06626-64C3-4683-90DE-BD87ADF64C29}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{75C06626-64C3-4683-90DE-BD87ADF64C29}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{75C06626-64C3-4683-90DE-BD87ADF64C29}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{75C06626-64C3-4683-90DE-BD87ADF64C29}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{75C06626-64C3-4683-90DE-BD87ADF64C29}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -460,7 +453,6 @@ Global
 		{75C06626-64C3-4683-90DE-BD87ADF64C29}.SERVER|x86.ActiveCfg = Release|Any CPU
 		{75C06626-64C3-4683-90DE-BD87ADF64C29}.SERVER|x86.Build.0 = Release|Any CPU
 		{EFE36109-BE2C-49CE-82DF-5649F61898C0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EFE36109-BE2C-49CE-82DF-5649F61898C0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EFE36109-BE2C-49CE-82DF-5649F61898C0}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{EFE36109-BE2C-49CE-82DF-5649F61898C0}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{EFE36109-BE2C-49CE-82DF-5649F61898C0}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -507,7 +499,6 @@ Global
 		{122612CC-F070-472E-877C-4DC9D86275E5}.SERVER|x86.ActiveCfg = Release|Any CPU
 		{122612CC-F070-472E-877C-4DC9D86275E5}.SERVER|x86.Build.0 = Release|Any CPU
 		{E1774431-C670-49CE-A879-63D3D1599F1A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E1774431-C670-49CE-A879-63D3D1599F1A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E1774431-C670-49CE-A879-63D3D1599F1A}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{E1774431-C670-49CE-A879-63D3D1599F1A}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{E1774431-C670-49CE-A879-63D3D1599F1A}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -577,7 +568,6 @@ Global
 		{B407D46B-4515-48B0-AA2A-E362DA4345AC}.SERVER|x86.ActiveCfg = Release|Any CPU
 		{B407D46B-4515-48B0-AA2A-E362DA4345AC}.SERVER|x86.Build.0 = Release|Any CPU
 		{F92E896D-3E0E-41B3-90AE-FEDAD994AD1D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F92E896D-3E0E-41B3-90AE-FEDAD994AD1D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F92E896D-3E0E-41B3-90AE-FEDAD994AD1D}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{F92E896D-3E0E-41B3-90AE-FEDAD994AD1D}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{F92E896D-3E0E-41B3-90AE-FEDAD994AD1D}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -717,7 +707,6 @@ Global
 		{583B2E53-7117-4BA3-8B11-4E568AE7157C}.SERVER|x86.ActiveCfg = Debug|Any CPU
 		{583B2E53-7117-4BA3-8B11-4E568AE7157C}.SERVER|x86.Build.0 = Debug|Any CPU
 		{0ACD233F-DC92-4456-818C-F6841ADDD9CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0ACD233F-DC92-4456-818C-F6841ADDD9CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0ACD233F-DC92-4456-818C-F6841ADDD9CB}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{0ACD233F-DC92-4456-818C-F6841ADDD9CB}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{0ACD233F-DC92-4456-818C-F6841ADDD9CB}.Debug|x64.ActiveCfg = Debug|Any CPU
